Create README.md
The Sustainable  Energy Prediction App is a powerful tool designed to forecast future energy production using linear regression models. These models are trained on 30 years of historical data for both solar and wind energy. Leveraging the simplicity and interpretability of linear regression and the capabilities of the Scikit-Learn library, the app provides valuable insights into the expected energy output for user-specified future years.